simplify generic type classes
