+++ src/main/java/com/mercateo/jsonschema/generictype/GenericParameterizedType.kt
@@ -22,17 +22,6 @@ internal class GenericParameterizedType<T>(type: ParameterizedType, rawType: Cla
             return GenericType.of(actualTypeArguments[0], rawType)
         }
 
-    override val superType: GenericType<Any>?
-        get() {
-            val superclass = rawType.superclass as Class<out Any>?
-            return if (superclass != null) {
-                val exactSuperType: Type = GenericTypeReflector.getExactSuperType(type,
-                        superclass)
-                GenericType.of(exactSuperType, superclass)
-            } else
-                null
-        }
-
     override fun toString(): String {
         return type.toString()
     }
